+++ gfn_divisor/GFNDivisorTester.cpp
@@ -751,7 +751,7 @@ void  GFNDivisorTester::CheckRedc(mp_limb_t *xp, uint32_t xn, uint32_t b, uint32
       {         
          ip_App->WriteToConsole(COT_OTHER, "REDC ERROR: %u^2^%u mod %" PRIu64"*2^%u+1\n", b, m, k, n);
          
-         FatalError("limb %u=%" PRIu64", xp[%u]=%" PRIu64"\n", i, (uint64_t) mpz_getlimbn(B, i), i, xp[i]);
+         FatalError("limb %u=%" PRIu64", xp[%u]=%" PRIu64"\n", i, (uint64_t) mpz_getlimbn(B, i), i, (uint64_t) xp[i]);
       }
 
    mpz_clear(N);
